Split server into lib/server and server.js to separate responsibilities, allowing injecting custom twitch chat clients. Temp fix to test.js. A true fix would create an entire mocked twitch chat client rather than just bits and pieces, but that can happen in the future.
