+++ lib/twitch_chat.js
@@ -229,6 +229,10 @@ class TwitchChatClient {
 	 * @param {*} channels
 	 */
 	heartbeat(channels) {
+		if (!this.connected) {
+			return;
+		}
+
 		const chatClient = this.chatClient;
 
 		// If we're not in any of these channels, join them.
